Setup to have merged russh/libssh impls.
